fix type

+++ mace/modules/radial.py
@@ -192,8 +192,8 @@ class ZBLBasis(torch.nn.Module):
         node_atomic_numbers = atomic_numbers[torch.argmax(node_attrs, dim=1)].unsqueeze(
             -1
         )
-        Z_u = node_atomic_numbers[sender]
-        Z_v = node_atomic_numbers[receiver]
+        Z_u = node_atomic_numbers[sender].to(torch.int64)
+        Z_v = node_atomic_numbers[receiver].to(torch.int64)
         a = (
             self.a_prefactor
             * 0.529
@@ -309,8 +309,8 @@ class SoftTransform(torch.nn.Module):
         node_atomic_numbers = atomic_numbers[torch.argmax(node_attrs, dim=1)].unsqueeze(
             -1
         )
-        Z_u = node_atomic_numbers[sender]
-        Z_v = node_atomic_numbers[receiver]
+        Z_u = node_atomic_numbers[sender].to(torch.int64)
+        Z_v = node_atomic_numbers[receiver].to(torch.int64)
         r_0 = (self.covalent_radii[Z_u] + self.covalent_radii[Z_v]) / 4
         y = (
             x
